Fix incorrect shift operator binding

+++ lucene/core/src/java/org/apache/lucene/util/SparseFixedBitSet.java
@@ -412,7 +412,7 @@ public class SparseFixedBitSet extends BitSet {
       long[] newBits = ArrayUtil.copyOfSubArray(bits, 0, nonZeroLongCount);
       this.bits[i4096] = newBits;
       // we may slightly overestimate size here, but keep it cheap
-      this.ramBytesUsed += SINGLE_ELEMENT_ARRAY_BYTES_USED + (newBits.length - 1) << 3;
+      this.ramBytesUsed += SINGLE_ELEMENT_ARRAY_BYTES_USED + ((long) newBits.length - 1 << 3);
       this.nonZeroLongCount += nonZeroLongCount;
       return;
     }
